refactor(config): extract Profile dataclasses from credentials.py into profiles.py (#721)
Phase 2 of 2 — finishes #721. Move the 13 typed *Profile dataclasses + the
ProfileConfig union out of credentials.py into a new drt/config/profiles.py;
credentials.py keeps env-resolution / secret-loading / load-save dispatch and
re-exports the profiles with an explicit __all__ (25-name public surface) so
every `from drt.config.credentials import XProfile` keeps working.

Mechanical move only — no dataclass field/default/behavior change. Each profile
keeps its own describe(): DescribableConfig (from PR #1's base.py) is a pydantic
BaseModel and can't be a base for stdlib @dataclasses without converting them to
pydantic (a behavior change), so the describe() dedup is intentionally skipped
this phase.

Verified:
- Whole-package `python -m mypy drt/` clean — 0 no_implicit_reexport errors across
  151 files (the explicit __all__ is what makes the re-exports count under strict
  mypy). Only remaining error is the pre-existing credentials.py:175 [no-any-return]
  in load_snowflake_private_key (function byte-identical to main; line shifted from
  368 only because the profiles above it moved out).
- Structural parity: snapshotted every profile's dataclasses.fields()
  (names/types/defaults/factories) + describe() before vs after — 0 diffs across
  14 entries, describe() identical 13/13.
- ruff clean; full suite green (1876 passed; the 6 failures are the pre-existing
  Windows test_sqlite.py os.unlink flake).

Closes #721
